basic products container and component created
